# Optimized Eytzinger layout & memory prefetch