--- a/benches/search_comparison.rs
+++ b/benches/search_comparison.rs
@@ -38,7 +38,8 @@ where
     let plot_config = PlotConfiguration::default().summary_scale(AxisScale::Logarithmic);
 
     let sizes = [
-        8, 16, 32, 64, 128, 256, 512, 1024, 2048, 4069, 8192, 16384, 32768, 65536,
+        8, 16, 32, 64, 128, 256, 512, 1024, 2048, 4069, 8192, 16384, 32768, 65536, 1048576,
+        10485760,
     ];
 
     {