When using prefetch, the search loop became heavily CPU-backend bound. Basically, prefetch offset computation logic is competing with search logic itself for ALU.

This PR contains 2 optimizations of the search loop that allows us to do the same work using fewer arithmetics and brings 5-30% of the performance for `OrderedCollection` depending on the type and array size.

## Modified Eytzinger layout

The original navigation rules for Eytzinger are `2i + 1`/`2i + 2`. But if we put the root element in the 1st element in the array instead of 0th, iteration rules becomes `2i`/`2i+1`. It helps in several ways:

1. simpler navigation arithmetics
1. simpler prefetch offset – it is just half of a cache line
1. greatly simpler branchless index decode arithmetics

## Partially masked prefetch

The idea of a masked prefetch was initially described in [1]. The motivation is the following

> With some processors (the AMD Phenom II, Intel Celeron J1900, and Intel E3-1230 are examples) the code that does explicit prefetching is much slower, especially for smaller values of n. This seems to be due to executing prefetch instructions that attempt to prefetch data outside the input array. Although the documentation for the __builtin_prefetch() instruction specifically allows prefetching invalid addresses, this seems to cause a significant performance hit on these processors. On such architectures, a workaround is to use a bit-mask to ensure that we never prefetch an address outside the input array

Indeed the described effect was detectable on every CPU I've run the tests on. However, after the enormous amount of days spent on PMU analysis, I tend to believe the explanation is not correct.

The prefetch logic is approximately doubling the amount of arithmetic that needs to be performed per each loop iteration. For large arrays when we're memory bound it pays off, but for smaller `n` it hurts the performance. There is clear evidence (Intel VTune/perf) that for small arrays the loop is backend-bound with high port 0, 1, and 5 utilization (which are connected to ALU on Intel) (see. [Understanding CPU port contention](https://easyperf.net/blog/2018/03/21/port-contention)).

Now the reason masking helps, I believe, is not because of some microarchitecture specifics of some CPUs, but because on the last iteration of a search loop masked approach allows to prefetch first elements of the array helping subsequent searches to perform faster. Essentially, it is a cheap version of `offset % size`

The optimization implemented is moving offset out of masking.

```
current:   prefetch = base + (multiplier * i + offset) & mask
optimized: prefetch = base + offset + (multiplier * i) & mask
```

It allows us to make an offset part of the base address and move it out of the loop. Although the optimized approach formally produces slightly different addresses this optimization works quite nicely with the previous one, because after Eytzinger layout optimization offset is always half a cache line. So we're always prefetching addresses not further than half a cache line from the original approach.

I also added some large array size benchmarks to test that performance doesn't get worse on workloads that doesn't fit in cache.

[1]: [ARRAY LAYOUTS FOR COMPARISON-BASED SEARCHING](https://arxiv.org/pdf/1509.05053v1.pdf)



<details>
  <summary>Criterion results (AWS c1.medium: Intel Xeon E5-2651 v2 @ 1.80GHz)</summary>

```
Search u8/ordsearch/8   time:   [23.503 ns 23.521 ns 23.539 ns]
                        change: [-16.089% -15.614% -14.965%] (p = 0.00 < 0.05)
                        Performance has improved.
Search u8/ordsearch/16  time:   [23.219 ns 23.242 ns 23.273 ns]
                        change: [-13.184% -12.386% -11.456%] (p = 0.00 < 0.05)
                        Performance has improved.
Search u8/ordsearch/32  time:   [25.123 ns 25.164 ns 25.207 ns]
                        change: [-11.313% -10.526% -9.3601%] (p = 0.00 < 0.05)
                        Performance has improved.
Search u8/ordsearch/64  time:   [26.657 ns 26.670 ns 26.685 ns]
                        change: [-14.855% -14.052% -13.202%] (p = 0.00 < 0.05)
                        Performance has improved.
Search u8/ordsearch/128 time:   [28.850 ns 28.870 ns 28.897 ns]
                        change: [-15.515% -14.757% -13.771%] (p = 0.00 < 0.05)
                        Performance has improved.
Search u8/ordsearch/256 time:   [31.479 ns 31.496 ns 31.516 ns]
                        change: [-18.131% -17.705% -16.973%] (p = 0.00 < 0.05)
                        Performance has improved.
Search u8/ordsearch/512 time:   [34.661 ns 34.680 ns 34.703 ns]
                        change: [-19.898% -19.542% -18.967%] (p = 0.00 < 0.05)
                        Performance has improved.
Search u8/ordsearch/1024
                        time:   [38.093 ns 38.184 ns 38.367 ns]
                        change: [-21.644% -21.185% -20.680%] (p = 0.00 < 0.05)
                        Performance has improved.
Search u8/ordsearch/2048
                        time:   [42.199 ns 42.232 ns 42.277 ns]
                        change: [-21.741% -21.400% -20.912%] (p = 0.00 < 0.05)
                        Performance has improved.
Search u8/ordsearch/4069
                        time:   [87.869 ns 87.925 ns 87.979 ns]
                        change: [+49.039% +49.385% +49.683%] (p = 0.00 < 0.05)
                        Performance has regressed.
Search u8/ordsearch/8192
                        time:   [51.234 ns 51.267 ns 51.305 ns]
                        change: [-19.556% -19.408% -19.282%] (p = 0.00 < 0.05)
                        Performance has improved.
Search u8/ordsearch/16384
                        time:   [55.749 ns 55.794 ns 55.845 ns]
                        change: [-19.294% -19.077% -18.719%] (p = 0.00 < 0.05)
                        Performance has improved.
Search u8/ordsearch/32768
                        time:   [60.393 ns 60.420 ns 60.446 ns]
                        change: [-18.621% -18.500% -18.395%] (p = 0.00 < 0.05)
                        Performance has improved.
Search u8/ordsearch/65536
                        time:   [64.436 ns 64.479 ns 64.521 ns]
                        change: [-18.751% -18.609% -18.437%] (p = 0.00 < 0.05)
                        Performance has improved.
Search (with duplicates) u8/ordsearch/8
                        time:   [19.105 ns 19.126 ns 19.150 ns]
                        change: [-5.4435% -4.4330% -3.5011%] (p = 0.00 < 0.05)
                        Performance has improved.
Search (with duplicates) u8/ordsearch/16
                        time:   [21.618 ns 22.583 ns 23.681 ns]
                        change: [-6.5071% -3.6402% -0.6330%] (p = 0.01 < 0.05)
                        Change within noise threshold.
Search (with duplicates) u8/ordsearch/32
                        time:   [23.742 ns 23.758 ns 23.778 ns]
                        change: [-12.656% -11.565% -10.500%] (p = 0.00 < 0.05)
                        Performance has improved.
Search (with duplicates) u8/ordsearch/64
                        time:   [26.071 ns 26.137 ns 26.275 ns]
                        change: [-14.801% -14.048% -13.141%] (p = 0.00 < 0.05)
                        Performance has improved.
Search (with duplicates) u8/ordsearch/128
                        time:   [28.555 ns 28.578 ns 28.608 ns]
                        change: [-18.574% -17.865% -16.971%] (p = 0.00 < 0.05)
                        Performance has improved.
Search (with duplicates) u8/ordsearch/256
                        time:   [31.461 ns 31.476 ns 31.494 ns]
                        change: [-24.880% -24.384% -23.773%] (p = 0.00 < 0.05)
                        Performance has improved.
Search (with duplicates) u8/ordsearch/512
                        time:   [34.645 ns 34.673 ns 34.708 ns]
                        change: [-33.630% -33.133% -32.397%] (p = 0.00 < 0.05)
                        Performance has improved.
Search (with duplicates) u8/ordsearch/1024
                        time:   [38.087 ns 38.135 ns 38.213 ns]
                        change: [-43.187% -42.786% -42.323%] (p = 0.00 < 0.05)
                        Performance has improved.
Search (with duplicates) u8/ordsearch/2048
                        time:   [42.213 ns 42.231 ns 42.249 ns]
                        change: [-23.614% -22.205% -21.299%] (p = 0.00 < 0.05)
                        Performance has improved.
Search (with duplicates) u8/ordsearch/4069
                        time:   [87.734 ns 87.781 ns 87.829 ns]
                        change: [+48.902% +49.264% +49.683%] (p = 0.00 < 0.05)
                        Performance has regressed.
Search (with duplicates) u8/ordsearch/8192
                        time:   [51.130 ns 51.151 ns 51.171 ns]
                        change: [-19.684% -19.575% -19.467%] (p = 0.00 < 0.05)
                        Performance has improved.
Search (with duplicates) u8/ordsearch/16384
                        time:   [55.654 ns 55.682 ns 55.713 ns]
                        change: [-19.424% -19.258% -19.077%] (p = 0.00 < 0.05)
                        Performance has improved.
Search (with duplicates) u8/ordsearch/32768
                        time:   [60.351 ns 60.374 ns 60.399 ns]
                        change: [-18.547% -18.432% -18.298%] (p = 0.00 < 0.05)
                        Performance has improved.
Search (with duplicates) u8/ordsearch/65536
                        time:   [64.457 ns 64.503 ns 64.564 ns]
                        change: [-18.732% -18.475% -18.265%] (p = 0.00 < 0.05)
                        Performance has improved.
Search u16/ordsearch/8  time:   [23.344 ns 23.369 ns 23.398 ns]
                        change: [-13.889% -13.157% -12.274%] (p = 0.00 < 0.05)
                        Performance has improved.
Search u16/ordsearch/16 time:   [23.239 ns 23.249 ns 23.260 ns]
                        change: [-13.005% -12.249% -11.346%] (p = 0.00 < 0.05)
                        Performance has improved.
Search u16/ordsearch/32 time:   [25.009 ns 25.034 ns 25.060 ns]
                        change: [-11.576% -11.035% -10.324%] (p = 0.00 < 0.05)
                        Performance has improved.
Search u16/ordsearch/64 time:   [26.804 ns 26.829 ns 26.859 ns]
                        change: [-22.114% -19.758% -17.401%] (p = 0.00 < 0.05)
                        Performance has improved.
Search u16/ordsearch/128
                        time:   [28.964 ns 28.984 ns 29.005 ns]
                        change: [-15.604% -14.900% -14.087%] (p = 0.00 < 0.05)
                        Performance has improved.
Search u16/ordsearch/256
                        time:   [31.678 ns 31.699 ns 31.729 ns]
                        change: [-18.579% -17.962% -17.152%] (p = 0.00 < 0.05)
                        Performance has improved.
Search u16/ordsearch/512
                        time:   [34.922 ns 34.940 ns 34.964 ns]
                        change: [-19.774% -19.245% -18.581%] (p = 0.00 < 0.05)
                        Performance has improved.
Search u16/ordsearch/1024
                        time:   [38.697 ns 38.730 ns 38.779 ns]
                        change: [-20.148% -19.642% -18.694%] (p = 0.00 < 0.05)
                        Performance has improved.
Search u16/ordsearch/2048
                        time:   [43.072 ns 43.086 ns 43.102 ns]
                        change: [-51.495% -51.098% -50.559%] (p = 0.00 < 0.05)
                        Performance has improved.
Search u16/ordsearch/4069
                        time:   [47.747 ns 47.760 ns 47.774 ns]
                        change: [-19.074% -18.646% -18.052%] (p = 0.00 < 0.05)
                        Performance has improved.
Search u16/ordsearch/8192
                        time:   [52.015 ns 52.057 ns 52.117 ns]
                        change: [-18.345% -17.943% -17.483%] (p = 0.00 < 0.05)
                        Performance has improved.
Search u16/ordsearch/16384
                        time:   [56.869 ns 56.894 ns 56.923 ns]
                        change: [-17.359% -17.120% -16.810%] (p = 0.00 < 0.05)
                        Performance has improved.
Search u16/ordsearch/32768
                        time:   [61.732 ns 61.785 ns 61.856 ns]
                        change: [-16.698% -16.539% -16.405%] (p = 0.00 < 0.05)
                        Performance has improved.
Search u16/ordsearch/65536
                        time:   [66.928 ns 66.964 ns 67.000 ns]
                        change: [-15.343% -15.248% -15.160%] (p = 0.00 < 0.05)
                        Performance has improved.
Search (with duplicates) u16/ordsearch/8
                        time:   [18.971 ns 18.977 ns 18.984 ns]
                        change: [-4.8995% -3.8263% -2.5950%] (p = 0.00 < 0.05)
                        Performance has improved.
Search (with duplicates) u16/ordsearch/16
                        time:   [21.079 ns 21.087 ns 21.097 ns]
                        change: [-8.2803% -7.2503% -6.0791%] (p = 0.00 < 0.05)
                        Performance has improved.
Search (with duplicates) u16/ordsearch/32
                        time:   [24.012 ns 24.030 ns 24.055 ns]
                        change: [-11.206% -10.511% -9.8690%] (p = 0.00 < 0.05)
                        Performance has improved.
Search (with duplicates) u16/ordsearch/64
                        time:   [26.275 ns 26.289 ns 26.305 ns]
                        change: [-14.300% -13.610% -12.961%] (p = 0.00 < 0.05)
                        Performance has improved.
Search (with duplicates) u16/ordsearch/128
                        time:   [28.709 ns 28.728 ns 28.752 ns]
                        change: [-18.248% -17.483% -16.765%] (p = 0.00 < 0.05)
                        Performance has improved.
Search (with duplicates) u16/ordsearch/256
                        time:   [31.498 ns 31.522 ns 31.550 ns]
                        change: [-20.299% -19.861% -19.261%] (p = 0.00 < 0.05)
                        Performance has improved.
Search (with duplicates) u16/ordsearch/512
                        time:   [34.836 ns 34.852 ns 34.874 ns]
                        change: [-21.950% -21.426% -20.295%] (p = 0.00 < 0.05)
                        Performance has improved.
Search (with duplicates) u16/ordsearch/1024
                        time:   [38.644 ns 38.660 ns 38.680 ns]
                        change: [-22.412% -21.925% -21.279%] (p = 0.00 < 0.05)
                        Performance has improved.
Search (with duplicates) u16/ordsearch/2048
                        time:   [43.064 ns 43.097 ns 43.141 ns]
                        change: [-51.222% -50.738% -50.145%] (p = 0.00 < 0.05)
                        Performance has improved.
Search (with duplicates) u16/ordsearch/4069
                        time:   [49.101 ns 49.114 ns 49.127 ns]
                        change: [-41.458% -41.222% -40.862%] (p = 0.00 < 0.05)
                        Performance has improved.
Search (with duplicates) u16/ordsearch/8192
                        time:   [52.067 ns 52.104 ns 52.161 ns]
                        change: [-20.205% -19.781% -19.440%] (p = 0.00 < 0.05)
                        Performance has improved.
Search (with duplicates) u16/ordsearch/16384
                        time:   [56.455 ns 56.495 ns 56.533 ns]
                        change: [-19.662% -19.385% -18.969%] (p = 0.00 < 0.05)
                        Performance has improved.
Search (with duplicates) u16/ordsearch/32768
                        time:   [60.904 ns 60.958 ns 61.024 ns]
                        change: [-20.077% -19.266% -18.740%] (p = 0.00 < 0.05)
                        Performance has improved.
Search (with duplicates) u16/ordsearch/65536
                        time:   [65.540 ns 65.605 ns 65.683 ns]
                        change: [-21.171% -20.949% -20.715%] (p = 0.00 < 0.05)
                        Performance has improved.
Search u32/ordsearch/8  time:   [23.178 ns 23.189 ns 23.199 ns]
                        change: [-17.292% -16.240% -15.279%] (p = 0.00 < 0.05)
                        Performance has improved.
Search u32/ordsearch/16 time:   [23.698 ns 23.710 ns 23.723 ns]
                        change: [-12.924% -12.190% -11.411%] (p = 0.00 < 0.05)
                        Performance has improved.
Search u32/ordsearch/32 time:   [24.817 ns 24.834 ns 24.852 ns]
                        change: [-11.449% -10.651% -9.7005%] (p = 0.00 < 0.05)
                        Performance has improved.
Search u32/ordsearch/64 time:   [26.544 ns 26.563 ns 26.583 ns]
                        change: [-12.551% -11.708% -10.874%] (p = 0.00 < 0.05)
                        Performance has improved.
Search u32/ordsearch/128
                        time:   [28.801 ns 28.833 ns 28.871 ns]
                        change: [-14.124% -13.395% -12.590%] (p = 0.00 < 0.05)
                        Performance has improved.
Search u32/ordsearch/256
                        time:   [31.504 ns 31.517 ns 31.531 ns]
                        change: [-16.634% -15.870% -14.937%] (p = 0.00 < 0.05)
                        Performance has improved.
Search u32/ordsearch/512
                        time:   [34.491 ns 34.504 ns 34.517 ns]
                        change: [-19.045% -18.283% -17.515%] (p = 0.00 < 0.05)
                        Performance has improved.
Search u32/ordsearch/1024
                        time:   [38.517 ns 38.559 ns 38.611 ns]
                        change: [-18.390% -18.018% -17.410%] (p = 0.00 < 0.05)
                        Performance has improved.
Search u32/ordsearch/2048
                        time:   [43.816 ns 43.835 ns 43.857 ns]
                        change: [-51.358% -51.200% -51.026%] (p = 0.00 < 0.05)
                        Performance has improved.
Search u32/ordsearch/4069
                        time:   [47.491 ns 47.518 ns 47.551 ns]
                        change: [-18.131% -17.854% -17.658%] (p = 0.00 < 0.05)
                        Performance has improved.
Search u32/ordsearch/8192
                        time:   [52.106 ns 52.146 ns 52.185 ns]
                        change: [-16.814% -16.380% -16.114%] (p = 0.00 < 0.05)
                        Performance has improved.
Search u32/ordsearch/16384
                        time:   [56.739 ns 56.799 ns 56.852 ns]
                        change: [-16.685% -16.318% -16.025%] (p = 0.00 < 0.05)
                        Performance has improved.
Search u32/ordsearch/32768
                        time:   [61.133 ns 61.159 ns 61.188 ns]
                        change: [-16.325% -16.196% -16.069%] (p = 0.00 < 0.05)
                        Performance has improved.
Search u32/ordsearch/65536
                        time:   [66.084 ns 66.139 ns 66.186 ns]
                        change: [-15.993% -15.843% -15.682%] (p = 0.00 < 0.05)
                        Performance has improved.
Search (with duplicates) u32/ordsearch/8
                        time:   [18.811 ns 18.821 ns 18.833 ns]
                        change: [-4.3354% -3.0835% -1.7193%] (p = 0.00 < 0.05)
                        Performance has improved.
Search (with duplicates) u32/ordsearch/16
                        time:   [21.399 ns 21.413 ns 21.428 ns]
                        change: [-4.0088% -2.8282% -1.5666%] (p = 0.00 < 0.05)
                        Performance has improved.
Search (with duplicates) u32/ordsearch/32
                        time:   [23.765 ns 23.792 ns 23.819 ns]
                        change: [-10.838% -9.9325% -9.0563%] (p = 0.00 < 0.05)
                        Performance has improved.
Search (with duplicates) u32/ordsearch/64
                        time:   [25.988 ns 26.005 ns 26.023 ns]
…