@jonhoo 2024-01-06
This PR has reminded me that we should probably (in a separate PR) add something like
```rust
fn into_iter(self) -> impl Iterator<Item = T>;
impl<T> From<OrderedCollection<T>> for Vec<T> {}
```
so that people can get the items back out again afterwards. I don't think we want to guarantee anything about the order of the returned values, so (with the `[0]` hack), we'd implement the second one as
```rust
let last = self.items.len() - 1;
self.items.swap(0, last);
self.items.truncate(last);
// SAFETY: all items other than [0] were initialized, and [0] has been removed
let items: Vec<T> = unsafe { transmute(self.item) };
items.into_iter()
```
(and the first can be implemented in terms of the second)